fixed last 2 feature, bug in order history.
